small refactoring changes, remove unneeded dist folder, update gitignore
