+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules
-dist/*.map
+dist/
+data/
 tags
